More typescripting
